requirements.txt
aiohttp==3.13.2
fastmcp==2.13.3
mcp==1.21.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
